+++ K-Engine/src/K-Engine/Time.h
@@ -6,9 +6,9 @@ namespace K
 	class K_API Time
 	{
 	public:
-		static float startTime;
-		static float endTime;
-		static float deltaTimeValue;
+		static double startTime;
+		static double endTime;
+		static double deltaTimeValue;
 		static float deltaTime();
 	};
 }
